#9 refactoring to separate code
